move testAssemblies.py into singleAssemblies.py
